requirements.txt
Django==5.0.1
django-phonenumber-field==7.3.0
pillow==10.2.0
djangorestframework==3.14.0